- 0804 page navigation 기능 추가

+++ src/main/java/sptek/spdevteam/intern/common/domain/Pagination.java
@@ -33,7 +33,7 @@ public class Pagination {
         this.listCnt = listCnt;
 
         //전체 페이지수
-        this.pageCnt = (int) Math.ceil(listCnt/listSize);
+        this.pageCnt = (int)Math.ceil((double)listCnt/(double)listSize);
 
         //시작 페이지
         this.startPage = (range - 1) * rangeSize + 1 ;
